chore: split tests for isolation
